fix to work on netbsd/current

+++ ksyms_subr.h
@@ -4,7 +4,7 @@
 int ksyms_load(void);
 void ksyms_destroy(void);
 void ksyms_dump(void);
-const char *ksyms_lookupsym(uintptr_t);
+const char *ksyms_lookupsym(unsigned long);
 uintptr_t ksyms_lookup(const char *);
 
 #endif /* _KSYMS_SUBR_H_ */
